# tokio-sync: AtomicTask performance improved if task is current