--- a/tokio-sync/src/mpsc/chan.rs
+++ b/tokio-sync/src/mpsc/chan.rs
@@ -1,6 +1,6 @@
+use loom::futures::AtomicTask;
 use super::list;
 use futures::Poll;
-use futures::task::AtomicTask;
 
 use std::cell::UnsafeCell;
 use std::process;