@carllerche 2019-02-14
We should add a test that checks the `register_task(not_current)` issue.